fix: clear receiver mypy to zero and make receiver typecheck blocking
Fix real bugs found by mypy (heartbeat_missed/sdk_integrity alerts passed null
project_id and never dispatched; reused-variable and Optional-narrowing bugs;
widen ttl_hours to float) and scope only framework false positives. Scrub
internal planning terms from public comments.

+++ receiver/repositories/sessions.py
@@ -46,7 +46,7 @@ async def create_session(
     user_id: UUID,
     ip_address: Optional[str] = None,
     user_agent: Optional[str] = None,
-    ttl_hours: int = _DEFAULT_SESSION_TTL_HOURS,
+    ttl_hours: float = _DEFAULT_SESSION_TTL_HOURS,
 ) -> tuple[str, Session]:
     """Create a new session. Returns (raw_token, Session).
